refactor and create postman files
