+++ app/backend/handlers/tag-handler.js
@@ -2,7 +2,7 @@ var express = require('express');
 var router = express.Router();
 var TagService = require('../services/tag-service.js');
 
-router.get('/getAllTags', async (req, res) => {
+router.get('/getTags', async (req, res) => {
   const tags = await TagService.getAllTags();
   res.send({ tags });
 });
