+++ app/backend/handlers/user-handler.js
@@ -2,9 +2,14 @@ var express = require('express');
 var router = express.Router();
 var UserService = require('../services/user-service.js');
 
-router.get('/getAllUsernames', async (req, res) => {
-  const usernames = await UserService.getAllUsernames();
-  res.send({ usernames });
+router.get('/getUsers', async (req, res) => {
+  const users = await UserService.getUsers();
+  res.send({ users });
+});
+
+router.get('/getUserInfo/:userId', async (req, res) => {
+  const userInfo = await UserService.getUserInfo(req.params.userId);
+  res.send({ userInfo });
 });
 
 router.get('/getProfileInfo/:userId', async (req, res) => {
