+++ app/backend/services/recipe-service.js
@@ -1,13 +1,29 @@
-const RecipeRepo = require('../repositories/recipe-repository');
+var { getPool } = require('../config/db.js')
+var sql = require('mssql')
 
 async function getRecipes() {
-  var recipes = await RecipeRepo.getRecipes()
-  return recipes
+  try {
+    const pool = await getPool();
+    const recipes = await pool.request().query("SELECT * FROM Recipes");
+    return recipes.recordsets;
+  } catch (error) {
+    console.log(error);
+    return [];
+  }
 }
 
 async function getRecipeById(recipeId) {
-  var recipeById = await RecipeRepo.getRecipeById(recipeId)
-  return recipeById
+  try {
+    const pool = await getPool();
+    const recipeById = await pool.request()
+      .input('recipe_id', sql.Int, recipeId)
+      .query("SELECT * FROM Recipes WHERE Recipes.recipe_id=@recipe_id");
+    return recipeById.recordsets;
+
+  } catch (error) {
+    console.log(error);
+    return [];
+  }
 }
 
 module.exports = { getRecipes, getRecipeById };
