+++ app/backend/services/recipe-tag-service.js
@@ -1,8 +1,20 @@
-const RecipeTagRepo = require('../repositories/recipe-tag-repository');
+var { getPool } = require('../config/db.js')
+var sql = require('mssql');
 
-async function getTagsForRecipe(recipe_id) {
-  var tagsForRecipe = await RecipeTagRepo.getTagsForRecipe(recipe_id);
-  return tagsForRecipe;
+async function getTagsForRecipe(recipeId) {
+  try {
+    const pool = await getPool();
+    console.log("recipeId, ", recipeId);
+    const tagsForRecipe = await pool.request()
+      .input('recipe_id', sql.Int, recipeId)
+      .query("SELECT Tags.tag_name FROM Tags \
+                      INNER JOIN RecipeTags ON Tags.tag_id = RecipeTags.tag_id \
+                      WHERE RecipeTags.recipe_id=@recipe_id;");
+    return tagsForRecipe.recordsets;
+  } catch (error) {
+    console.log(error);
+    return [];
+  }
 }
 
 module.exports = { getTagsForRecipe };
