+++ app/backend/services/tag-service.js
@@ -1,8 +1,14 @@
-const TagRepo = require('../repositories/tag-repository');
+var { getPool } = require('../config/db.js')
 
 async function getAllTags() {
-  var tags = await TagRepo.getAllTags();
-  return tags;
+  try {
+    const pool = await getPool();
+    const tags = await pool.request().query("SELECT * FROM Tags");
+    return tags.recordsets;
+  } catch (error) {
+    console.log(error);
+    return [];
+  }
 }
 
 module.exports = { getAllTags };
